% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{mydata}
\alias{mydata}
\title{My Example Dataset}
\format{
My data
\describe{
\item{x}{numeric}
}
}
\source{
Generated for package example
}
\usage{
mydata
}
\description{
A dataset containing ...
}
\keyword{datasets}
