% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{myfunctions}
\alias{myfunctions}
\title{myfunctions}
\usage{
myfunctions(x, studentnumber)
}
\arguments{
\item{x}{numeric vector}

\item{studentnumber}{numeric value}
}
\value{
Add the \code{studentnumber} to  \code{x} and returns the mean and variance.
}
\description{
Add the student number to a vector x and calculates the variance and mean in C++.
}
